feat: Verify cached mini app files using SHA-256 hash
This is to prevent a cache poisining vulnerability. This works by saving an encrypted hash of the Mini App fils. If someone modifies the cached files for the Mini App (i.e. on a rooted device), then the hash verification will fail and the Mini App will be re-downloaded. It's also not possible for someone to modify the hash since it's encrypted - if they attempt to modify it then the decryption of the hash will fail.

I also needed to delete some unit tests related to SDK initialization - this is because the CachedMiniAppVerifier needs to use the AndroidKeyStore which isn't mocked by the test runner. However I think it's okay to not test the initialization code because part of it's purpose is to be the place where we can initialize things which aren't mockable.