docs: Reformat user guide into feature sections
Most of the content is still the same but was just rearranged. The only new content is the Mini App Features section which explains what each feature does. Also added a Table of Contents - this will be automatically generated by Jekyll when the documentation is published to Github Pages.

Also put some of the longer code examples inside of expandable sections.