fix: Issue with simultaneous custom permission requests
There was an issue when a Mini App tried to call MiniApp.requestCustomPermissions twice in a row without waiting for the success/failure response of the previous call. This is because the CustomPermissionBridgeDispatcher can only be initialized with a single CallbackObj. There wasn't any reason for only a single CustomPermissionBridgeDispatcher object to be used, so I moved it's instantiation inline when the custom permissions request is made.

Also refactored CustomPermissionBridgeDispatcher a little to remove the initCallbackObj as it was no longer nescessary