fix: Unable to launch mini app when device offline
This fixes a case where the the mini app is unable to be launched when the device is offline. The check being done on the manifest was failing when offline. We missed this case earlier because it only occurs when testing with multiple mini apps - we cache the manifest for the last launched mini app using ManifestApiCache